requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
python-dotenv
PyJWT
razorpay
